Restructure build; control preview hook work; cleanup & more...
This commit changes a ton.

Of note, I worked on the control preview hook quite a bit. It doesn't work
properly yet, but it's some good progress. Now a combined C/Rust program is
created and works just fine.

I also restructured the build system to share some common code between
control preview and main modules. There are refactorings I have yet to do,
but this is a good base.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>