Begin work on Restyle integration
Kanaya will rely on a simple fork of Restyle to function. This will make
the core Restyle into a DLL, from which Restyle functions may be accessed
by the Kanaya GUI, and a stub "restyle.exe" which can be used over the
command line just like the standalone project.

So, while Restyle itself just has "restyle.exe" implementing the entire
program, Kanaya has "restyle.exe" (a stub) and "kanaya.exe" (Kanaya GUI)
sharing "rstylkan.dll" (Restyle implementation).

restyle/src contains raw Restyle source code. This is kept raw so that the
Restyle submodule may be updated without issue.

restyle/kanayacpp contains Kanaya-specific code to be mixed in with the
Restyle source code. These files will just be dropped into the build source
folder as-is. As the name of the folder suggests, Kanaya extensions to
Restyle will be implemented in C++; Rust cannot be used in the first place
due to Restyle depending on Microsoft's C++ build tools.

restyle/srcpatches contains diffs to be applied to src files.

In my vision, there will also be a Cargo.toml and build.rs in the folder to
coordinate build. If Rust is ever to be integrated in some way, a
"kanayarust" folder shall be made. As for the stub, both the Cargo CC crate
and Rust itself are good options; we just need to call into one function
and let the DLL take control of everything from there.

Building the Restyle core (rstylkan.dll) will likely still require a slave
MSBuild process and utilise MSVC. The ultimate project which is built
will likely be generated in the Cargo OUT_DIR. This will likely be the
first thing to be built, because of the Kanaya GUI's dependence on services
from Restyle.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>